+++ src/geoplotnik/components/ids.py
@@ -15,9 +15,6 @@ TAS_DIAGRAM = f"{TAS}-diagram"
 TAS_DIAGRAM_CONTAINER = f"{TAS}-diagram-container"
 
 TAS_DIAGRAM_GROUPING_PARAMETER_DROPDOWN = f"{TAS}-grouping-parameter-dropdown"
-TAS_DIAGRAM_GROUPING_PARAMETER_SELECT_ALL_BUTTON = (
-    f"{TAS}-grouping-parameter-select-all-button"
-)
 
 TAS_DIAGRAM_X_AXIS_DROPDOWN = f"{TAS}-x-axis-dropdown"
 TAS_DIAGRAM_Y_AXIS_DROPDOWN = f"{TAS}-y-axis-dropdown"
